Add tests for Splinter backend
Adds an enum to select a backend, all existing tests must now
specify Backend::Sawtooth, while new tests with service IDs must
specify Backend::Splinter, otherwise the REST API will not handle
the request with a service_id query parameter.

This also adds a service_id argument to the functions which
generate the Grid objects to allow for this field to be set.

Also adds tests for objects with a service ID field set to some
value.

Signed-off-by: Shannyn Telander <telander@bitwise.io>